fixed authentiaction bugs

+++ src/main/java/dev/thural/quietspace/service/impls/AuthServiceImpl.java
@@ -1,26 +1,30 @@
 package dev.thural.quietspace.service.impls;
 
+import dev.thural.quietspace.entity.Role;
 import dev.thural.quietspace.entity.Token;
 import dev.thural.quietspace.entity.User;
-import dev.thural.quietspace.mapper.UserMapper;
-import dev.thural.quietspace.model.request.UserRequest;
+import dev.thural.quietspace.model.request.UserRegisterRequest;
 import dev.thural.quietspace.model.request.LoginRequest;
 import dev.thural.quietspace.model.response.AuthResponse;
+import dev.thural.quietspace.repository.RoleRepository;
 import dev.thural.quietspace.repository.TokenRepository;
 import dev.thural.quietspace.repository.UserRepository;
 import dev.thural.quietspace.security.JwtService;
 import dev.thural.quietspace.service.AuthService;
+import dev.thural.quietspace.utils.enums.RoleType;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashMap;
-import java.util.Optional;
+import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -28,57 +32,55 @@ import java.util.UUID;
 @Transactional
 public class AuthServiceImpl implements AuthService {
 
-    private final AuthenticationManager authenticationManager;
+
     private final PasswordEncoder passwordEncoder;
+    private final UserDetailsService userDetailsService;
     private final TokenRepository tokenRepository;
-    private final UserMapper userMapper;
     private final UserRepository userRepository;
     private final JwtService jwtService;
+    private final RoleRepository roleRepository;
 
 
     @Override
-    public AuthResponse register(UserRequest userRequest) {
+    public AuthResponse register(UserRegisterRequest user) {
+        String userPassword = user.getPassword();
+        user.setPassword(passwordEncoder.encode(userPassword));
 
-        String userPassword = userRequest.getPassword();
-        userRequest.setPassword(passwordEncoder.encode(userPassword));
-        User savedUser = userRepository.save(userMapper.userRequestToEntity(userRequest));
+        Role userRole = roleRepository.findByName(RoleType.USER.toString())
+                // todo - better exception handling
+                .orElseThrow(() -> new IllegalStateException("ROLE USER was not initiated"));
 
-        var auth = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(userRequest.getEmail(), userPassword)
+        User savedUser = userRepository.save(
+                User.builder()
+                        .username(user.getUsername())
+                        .firstname(user.getFirstname())
+                        .lastname(user.getLastname())
+                        .password(user.getPassword())
+                        .email(user.getEmail())
+                        .role(user.getRole())
+                        .roles(List.of(userRole))
+                        .accountLocked(false)
+                        .enabled(true) // TODO: rollback to false value after testing
+                        .build()
         );
-        var claims = new HashMap<String, Object>();
-        var user = (User) auth.getPrincipal();
-        claims.put("fullName", user.getFullName());
-        String token = jwtService.generateToken(claims, user);
 
+        String token = jwtService.generateToken(savedUser);
         String userId = savedUser.getId().toString();
         return new AuthResponse(UUID.randomUUID(), token, userId, "register success");
-
     }
+
 
     @Override
     public AuthResponse login(LoginRequest loginRequest) {
         String userEmail = loginRequest.getEmail();
-        String userPassword = loginRequest.getPassword();
+        User user = userRepository.findUserEntityByEmail(userEmail)
+                .orElseThrow(() -> new AuthenticationCredentialsNotFoundException("login failed"));
 
-
-        var auth = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(userEmail, userPassword)
-        );
-        var claims = new HashMap<String, Object>();
-        var user = ((User) auth.getPrincipal());
-        claims.put("fullName", user.getFullName());
-        String token = jwtService.generateToken(claims, user);
-
-        Optional<Token> existingToken = tokenRepository.getByEmail(userEmail);
-        if (existingToken.isPresent()) {
-            if (jwtService.isTokenValid(existingToken.get().getToken(), user))
-                token = existingToken.get().getToken();
-            tokenRepository.deleteByEmail(userEmail);
-        }
-
+        if (tokenRepository.existsByEmail(userEmail)) tokenRepository.deleteByEmail(userEmail);
+        String token = jwtService.generateToken(user);
         return new AuthResponse(UUID.randomUUID(), token, user.getId().toString(), "login success");
     }
+
 
     @Override
     public void logout(String authHeader) {
@@ -92,8 +94,19 @@ public class AuthServiceImpl implements AuthService {
     }
 
     @Override
-    public Authentication generateAuthentication(String email, String password) {
-        return new UsernamePasswordAuthenticationToken(email, password);
+    public Authentication generateAuthentication(String username, String password) {
+        UserDetails userDetails = userDetailsService.loadUserByUsername(username);
+
+        if (userDetails == null)
+            throw new BadCredentialsException("invalid username");
+
+        if (!passwordEncoder.matches(password, userDetails.getPassword()))
+            throw new BadCredentialsException("invalid password");
+
+        return new UsernamePasswordAuthenticationToken(
+                userDetails,
+                userDetails.getPassword(),
+                userDetails.getAuthorities());
     }
 
     @Override
@@ -105,7 +118,6 @@ public class AuthServiceImpl implements AuthService {
                 .email(email)
                 .build()
         );
-        SecurityContextHolder.clearContext();
     }
 
     @Override
